Complete second week

+++ Ionic/conFusion/www/js/controllers.js
@@ -166,13 +166,19 @@ angular.module('conFusion.controllers', [])
 
 }])
 
-.controller('DishDetailController', ['$scope', '$stateParams', 'menuFactory', 'baseURL', function ($scope, $stateParams, menuFactory, baseURL) {
+.controller('DishDetailController', ['$scope', '$stateParams', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicPopover', '$ionicModal', function ($scope, $stateParams, menuFactory, favoriteFactory, baseURL, $ionicPopover, $ionicModal) {
 
   $scope.baseURL = baseURL;
   $scope.dish = {};
   $scope.showDish = false;
   $scope.message = "Loading ...";
-
+  $scope.comment = {
+      rating: 5,
+      comment: "",
+      author: "",
+      date: ""
+  };
+  
   $scope.dish = menuFactory.getDishes()
     .get({
       id: parseInt($stateParams.id, 10)
@@ -186,6 +192,58 @@ angular.module('conFusion.controllers', [])
         $scope.message = "Error: " + response.status + " " + response.statusText;
       }
     );
+  
+  $ionicPopover.fromTemplateUrl('templates/dish-detail-popover.html', {
+    scope: $scope
+  }).then(function(popover) {
+    $scope.popover = popover;
+  });
+  
+  $scope.showPopover = function ($event) {
+    $scope.popover.show($event);
+  };
+  
+  $scope.addToFavorites = function () {
+    favoriteFactory.addToFavorites($scope.dish.id);
+    $scope.popover.hide();
+  };
+  
+  // Comment Modal Form
+  $ionicModal.fromTemplateUrl('templates/dish-comment.html', {
+    scope: $scope
+  }).then(function (modal) {
+    $scope.commentModal = modal;
+  });
+
+  $scope.closeComment = function () {
+    $scope.commentModal.hide();
+  };
+
+  $scope.showComment = function () {
+    $scope.commentModal.show();
+  };
+
+  $scope.doComment = function () {
+    
+    $scope.comment.date = new Date().toISOString();
+
+    $scope.dish.comments.push($scope.comment);
+    menuFactory.getDishes().update({
+      id: $scope.dish.id
+    }, $scope.dish);
+    
+    $scope.comment = {
+      rating: 5,
+      comment: "",
+      author: "",
+      date: ""
+    };
+    
+    $scope.closeComment();
+    $scope.popover.hide();
+  };
+  
+  
 }])
 
 .controller('DishCommentController', ['$scope', 'menuFactory', function ($scope, menuFactory) {
@@ -253,19 +311,31 @@ angular.module('conFusion.controllers', [])
 
 }])
 
-.controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
+.controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', '$ionicPopup', '$ionicLoading', '$timeout', 
+function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout) {
 
   $scope.baseURL = baseURL;
-  $scope.shouldShowDelete  =false;
+  $scope.shouldShowDelete = false;
+
+  $ionicLoading.show({
+    template: '<ion-spinner></ion-spinner> Loading...'
+  });
   
   $scope.favorites = favoriteFactory.getFavorites();
-  
+
   $scope.dishes = menuFactory.getDishes().query(
     function (response) {
       $scope.dishes = response;
+      $timeout(function(){
+        $ionicLoading.hide();
+      }, 1000);
     },
     function (response) {
       $scope.message = "Error: " + response.status + " " + response.statusText;
+      $timeout(function(){
+        $ionicLoading.hide();
+      }, 1000);
+
     });
   
   console.log($scope.dishes, $scope.favorites);
@@ -273,30 +343,38 @@ angular.module('conFusion.controllers', [])
   $scope.toggleDelete = function () {
     $scope.shouldShowDelete = !$scope.shouldShowDelete;
     console.log($scope.shouldShowDelete);
-  };
-  
+  }
+
   $scope.deleteFavorite = function (index) {
-    favoriteFactory.deleteFromFavorites(index);
+
+    var confirmPopup = $ionicPopup.confirm({
+      title: 'Confirm Delete',
+      template: 'Are you sure you want to delete this item?'
+    });
+
+    confirmPopup.then(function (res) {
+      if (res) {
+        console.log('Ok to delete');
+        favoriteFactory.deleteFromFavorites(index);
+      } else {
+        console.log('Canceled delete');
+      }
+    });
+    
     $scope.shouldShowDelete = false;
-  };
-}
+  }
+}])
 
 .filter('favoriteFilter', function () {
   return function (dishes, favorites) {
     var out = [];
-    
     for (var i = 0; i < favorites.length; i++) {
-      
       for (var j = 0; j < dishes.length; j++) {
-        
-        if (dishes[j].id === favorites[i].id) {
-          
+        if (dishes[j].id === favorites[i].id)
           out.push(dishes[j]);
-        }
       }
     }
-    
     return out;
+
   }
-})
-;
+});;
